refactor: move Preferences, delete unused classes
